+++ ms-clientes/src/test/java/isi/dan/msclientes/dao/ClienteRepositoryTest.java
@@ -66,7 +66,7 @@ public class ClienteRepositoryTest {
         cliente.setCuit("123");
         cliente.setCorreoElectronico("asa@asfs.com");
         cliente.setMaxObrasEnEjecucion(3);
-        cliente.setMaximoDescubierto(10000);
+        cliente.setMaximoDescubierto(BigDecimal.valueOf(10000));
         cliente.setNombre("prueba");
         //clienteRepository.save(cliente);
     }
